refactor: Use Event Emitter API for creating a server
Refactor the code to use the Event Emitter API for creating a server that listens to server requests and provides appropriate responses. This improves the code structure and readability.

The changes include:
- Importing the `http` module
- Creating a server using `http.createServer`
- Subscribing to the "request" event using `server.on`
- Sending a "Welcome" response to the client

This commit improves the implementation of the server by utilizing the Event Emitter API instead of the commented out code block.